+++ CHANGELOG.md
@@ -253,6 +253,7 @@
 * Added generalized "slide show" for scripted sequences
 * Dialogue backgrounds are now fit to rather than enveloping screen (intended behaviour, previous setting was an error)
 * Fixed dialogue reactivating hidden objects even if they were initially inactive
+* Added new FindDeepChildren to SceneUtils and deprecated semi-broken FindAllGameObjects
 * Added NavigationNodeEx script [Experimental]
 * Fixed ThemeableElement and NonThemeableElement being ignored on child components of complex elements [Untested]
 * Added MaxChargeFraction to ShieldParams [Untested]